@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <IsPackable>false</IsPackable>
     <AssemblyName>tanka.graphql.dev.allocations</AssemblyName>
     <RootNamespace>Tanka.GraphQL.Dev.Allocations</RootNamespace>
@@ -19,6 +19,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\benchmarks\graphql.benchmarks\graphql.benchmarks.csproj" />
     <ProjectReference Include="..\..\src\graphql.language\graphql.language.csproj" />
   </ItemGroup>
 
