@@ -1,15 +1,15 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFrameworks>net6.0</TargetFrameworks>
-    <AssemblyName>Tanka.GraphQL.Extensions.ApolloFederation</AssemblyName>
-    <RootNamespace>Tanka.GraphQL.Extensions.ApolloFederation</RootNamespace>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
+	<PropertyGroup>
+		<TargetFrameworks>net7.0</TargetFrameworks>
+		<AssemblyName>Tanka.GraphQL.Extensions.ApolloFederation</AssemblyName>
+		<RootNamespace>Tanka.GraphQL.Extensions.ApolloFederation</RootNamespace>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\graphql.language\graphql.language.csproj" />
-    <ProjectReference Include="..\graphql\graphql.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\graphql.language\graphql.language.csproj" />
+		<ProjectReference Include="..\GraphQL\GraphQL.csproj" />
+	</ItemGroup>
 
 </Project>
