@@ -1,13 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net6.0</TargetFrameworks>
+    <TargetFrameworks>net7.0</TargetFrameworks>
     <RootNamespace>Tanka.GraphQL.Extensions.Tracing</RootNamespace>
     <AssemblyName>Tanka.GraphQL.Extensions.Tracing</AssemblyName>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\graphql\graphql.csproj" />
+    <ProjectReference Include="..\GraphQL\GraphQL.csproj" />
+    <ProjectReference Include="..\graphql\GraphQL.Legacy.csproj" />
   </ItemGroup>
 
 </Project>
