@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
-    <AssemblyName>tanka.graphql.language</AssemblyName>
+    <TargetFrameworks>net7.0;netstandard2.0</TargetFrameworks>
+	  <AssemblyName>tanka.graphql.language</AssemblyName>
     <RootNamespace>Tanka.GraphQL.Language</RootNamespace>
     <Nullable>enable</Nullable>
     <IsPackable>true</IsPackable>
@@ -13,8 +13,12 @@
     <PackageReference Include="Microsoft.Bcl.HashCode" Version="1.1.1" />
   </ItemGroup>
 
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0' ">
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net7.0' ">
     <Compile Remove="Shims/**" />
     <Compile Remove="System.Diagnostics.CodeAnalysis/**" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="System.Memory" Version="4.5.5" />
   </ItemGroup>
 </Project>
