@@ -6,13 +6,13 @@ namespace Tanka.GraphQL.Samples.Chat.Data;
 
 public interface IChatResolverService
 {
-    ValueTask<IResolverResult> GetMessagesAsync(IResolverContext context);
+    ValueTask GetMessagesAsync(ResolverContext context);
 
-    ValueTask<IResolverResult> AddMessageAsync(IResolverContext context);
+    ValueTask AddMessageAsync(ResolverContext context);
 
-    ValueTask<IResolverResult> EditMessageAsync(IResolverContext context);
+    ValueTask EditMessageAsync(ResolverContext context);
 
-    ValueTask<ISubscriberResult> StreamMessagesAsync(IResolverContext context, CancellationToken cancellationToken);
+    ValueTask StreamMessagesAsync(SubscriberContext context, CancellationToken cancellationToken);
 
-    ValueTask<IResolverResult> ResolveMessageAsync(IResolverContext context);
+    ValueTask ResolveMessageAsync(ResolverContext context);
 }
