@@ -1,18 +1,30 @@
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
+using System.Text.Json.Serialization;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Http.Json;
+using Microsoft.Extensions.DependencyInjection;
+using Tanka.GraphQL.Samples.Chat.Data;
+using Tanka.GraphQL.Server;
 
-namespace Tanka.GraphQL.Samples.Chat.Web;
+var builder = WebApplication.CreateBuilder(args);
 
-public class Program
+builder.Services.Configure<JsonOptions>(json =>
 {
-    public static void Main(string[] args)
-    {
-        CreateWebHostBuilder(args).Build().Run();
-    }
+    json.SerializerOptions.DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull;
+});
+// configure services
+builder.AddTankaGraphQL3()
+    .AddSchema("chat", options => { options.Configure(schema => schema.AddChat()); })
+    .AddHttp()
+    .AddWebSockets();
 
-    public static IWebHostBuilder CreateWebHostBuilder(string[] args)
-    {
-        return WebHost.CreateDefaultBuilder(args)
-            .UseStartup<Startup>();
-    }
-}
+builder.Services.AddSingleton<IChatResolverService, ChatResolverService>();
+builder.Services.AddSingleton<IChat, Chat>();
+
+var app = builder.Build();
+
+app.UseWebSockets();
+
+// this uses the default pipeline
+app.MapTankaGraphQL3("/graphql", "chat");
+
+app.Run();
