@@ -1,6 +1,8 @@
-﻿namespace Tanka.GraphQL.Extensions.ApolloFederation;
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace Tanka.GraphQL.Extensions.ApolloFederation;
 
 public interface IReferenceResolversMap
 {
-    bool TryGetReferenceResolver(string type, out ResolveReference resolveReference);
+    bool TryGetReferenceResolver(string type, [NotNullWhen(true)]out ResolveReference? resolveReference);
 }
