@@ -6,6 +6,6 @@ internal static class TimeSpanExtensions
 {
     public static long TotalNanoSeconds(this TimeSpan timeSpan)
     {
-        return (long)(timeSpan.TotalMilliseconds * 1000 * 1000);
+        return (long)timeSpan.TotalNanoseconds;
     }
 }
