@@ -26,7 +26,7 @@ public ref struct Lexer
 
     public bool IsExponential;
 
-    public Lexer(in ReadOnlySpan<byte> span)
+    public Lexer(ReadOnlySpan<byte> span)
     {
         _reader = new SpanReader(span);
         _currentLineStart = 0;
@@ -37,12 +37,12 @@ public ref struct Lexer
         IsExponential = false;
     }
 
-    public static Lexer Create(in ReadOnlySpan<byte> span)
+    public static Lexer Create(ReadOnlySpan<byte> span)
     {
         return new Lexer(span);
     }
 
-    public static Lexer Create(in string data)
+    public static Lexer Create(string data)
     {
         return Create(Encoding.UTF8.GetBytes(data));
     }
@@ -69,6 +69,12 @@ public ref struct Lexer
                 ReadStringValue();
                 return true;
             }
+
+            /*if (_reader.IsNext(Constants.EscapedQuote))
+            {
+                ReadStringValue();
+                return true;
+            }*/
 
             if (Constants.IsPunctuator(code))
             {
@@ -301,6 +307,8 @@ public ref struct Lexer
                     _reader.Advance();
                     StartNewLine();
                     break;
+                case Constants.Backslash:
+
                 case Constants.Return:
                     _reader.Advance();
                     break;
